Remove vertical bar and snap to window ability (#3729)
* Remove vertical bar and snap to window ability

* Missed a string
