+++ tools/DevInsights/DevHome.DevInsights/Strings/en-us/Resources.resw
@@ -1,17 +1,17 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <root>
-  <!--
-    Microsoft ResX Schema
-
+  <!-- 
+    Microsoft ResX Schema 
+    
     Version 2.0
-
-    The primary goals of this format is to allow a simple XML format
-    that is mostly human readable. The generation and parsing of the
-    various data types are done through the TypeConverter classes
+    
+    The primary goals of this format is to allow a simple XML format 
+    that is mostly human readable. The generation and parsing of the 
+    various data types are done through the TypeConverter classes 
     associated with the data types.
-
+    
     Example:
-
+    
     ... ado.net/XML headers & schema ...
     <resheader name="resmimetype">text/microsoft-resx</resheader>
     <resheader name="version">2.0</resheader>
@@ -26,36 +26,36 @@
         <value>[base64 mime encoded string representing a byte array form of the .NET Framework object]</value>
         <comment>This is a comment</comment>
     </data>
-
-    There are any number of "resheader" rows that contain simple
+                
+    There are any number of "resheader" rows that contain simple 
     name/value pairs.
-
-    Each data row contains a name, and value. The row also contains a
-    type or mimetype. Type corresponds to a .NET class that support
-    text/value conversion through the TypeConverter architecture.
-    Classes that don't support this are serialized and stored with the
+    
+    Each data row contains a name, and value. The row also contains a 
+    type or mimetype. Type corresponds to a .NET class that support 
+    text/value conversion through the TypeConverter architecture. 
+    Classes that don't support this are serialized and stored with the 
     mimetype set.
-
-    The mimetype is used for serialized objects, and tells the
-    ResXResourceReader how to depersist the object. This is currently not
+    
+    The mimetype is used for serialized objects, and tells the 
+    ResXResourceReader how to depersist the object. This is currently not 
     extensible. For a given mimetype the value must be set accordingly:
-
-    Note - application/x-microsoft.net.object.binary.base64 is the format
-    that the ResXResourceWriter will generate, however the reader can
+    
+    Note - application/x-microsoft.net.object.binary.base64 is the format 
+    that the ResXResourceWriter will generate, however the reader can 
     read any of the formats listed below.
-
+    
     mimetype: application/x-microsoft.net.object.binary.base64
-    value   : The object must be serialized with
+    value   : The object must be serialized with 
             : System.Runtime.Serialization.Formatters.Binary.BinaryFormatter
             : and then encoded with base64 encoding.
-
+    
     mimetype: application/x-microsoft.net.object.soap.base64
-    value   : The object must be serialized with
+    value   : The object must be serialized with 
             : System.Runtime.Serialization.Formatters.Soap.SoapFormatter
             : and then encoded with base64 encoding.
 
     mimetype: application/x-microsoft.net.object.bytearray.base64
-    value   : The object must be serialized into a byte array
+    value   : The object must be serialized into a byte array 
             : using a System.ComponentModel.TypeConverter
             : and then encoded with base64 encoding.
     -->
@@ -133,10 +133,6 @@
     <value>Unregister</value>
     <comment>The label for Unregister menu flyout for an icon in the toolbar. This will cause the program to no longer be registered with the bar.</comment>
   </data>
-  <data name="RotateLayoutButton.ToolTipService.ToolTip" xml:space="preserve">
-    <value>Switch between horizontal and vertical layout</value>
-    <comment>The tooltip for a button which toggles the orientation of the bar between horizontal and vertical</comment>
-  </data>
   <data name="DissociateButton.ToolTipService.ToolTip" xml:space="preserve">
     <value>Dissociate so that Dev Insights is not tracking any app</value>
     <comment>The tooltip for a button which causes Dev Insights to no longer track any app</comment>
@@ -144,10 +140,6 @@
   <data name="SwitchToLargeLayoutToolTip" xml:space="preserve">
     <value>Expand the large content panel</value>
     <comment>The tooltip for a button which expands the large content panel.</comment>
-  </data>
-  <data name="ReSnapButton.ToolTipService.ToolTip" xml:space="preserve">
-    <value>Snap the bar to the current foreground window</value>
-    <comment>The tooltip for a button that causes the bar to snap to the current foreground window.</comment>
   </data>
   <data name="SettingsButton.ToolTipService.ToolTip" xml:space="preserve">
     <value>Settings</value>
@@ -857,17 +849,9 @@
     <value>Select a process to investigate</value>
     <comment>Describes using a button in Dev Insights' UX to investigate another process running on the system</comment>
   </data>
-  <data name="SnapToolTip" xml:space="preserve">
-    <value>Attach the diagnostic bar to the target application</value>
-    <comment>The tooltip text for a button that attaches the bar to the target application.</comment>
-  </data>
   <data name="Title.Text" xml:space="preserve">
     <value>Dev Insights</value>
     <comment>{Locked} The product name</comment>
-  </data>
-  <data name="VerticalCloseButton.ToolTipService.ToolTip" xml:space="preserve">
-    <value>Close</value>
-    <comment>Describes a button that will close Dev Insights's window</comment>
   </data>
   <data name="CpuArchitecture_ProcessOnMachine" xml:space="preserve">
     <value>{0} process running on {1} machine</value>
@@ -904,10 +888,6 @@
   <data name="DetachMenuItem.Text" xml:space="preserve">
     <value>Detach</value>
     <comment>Menu Item that allows the user to detach from the current process</comment>
-  </data>
-  <data name="UnsnapToolTip" xml:space="preserve">
-    <value>Detach the diagnostic bar from the target application</value>
-    <comment>The tooltip text for a button that detaches the bar from the target application.</comment>
   </data>
   <data name="WERDumpColumn.Header" xml:space="preserve">
     <value>Dump file</value>
@@ -1220,10 +1200,6 @@ Dump File creation time: {1}</value>
   </data>
   <data name="ExpandCollapseLayoutButton.[using:Microsoft.UI.Xaml.Automation]AutomationProperties.Name" xml:space="preserve">
     <value>Expand or Collapse</value>
-    <comment>Name of the button expands or collapses the UI.</comment>
-  </data>
-  <data name="RotateLayoutButton.[using:Microsoft.UI.Xaml.Automation]AutomationProperties.Name" xml:space="preserve">
-    <value>Rotate Layout</value>
     <comment>Name of the button expands or collapses the UI.</comment>
   </data>
   <data name="NavLinksList.[using:Microsoft.UI.Xaml.Automation]AutomationProperties.Name" xml:space="preserve">
