+++ tools/DevInsights/DevHome.DevInsights/ViewModels/BarWindowViewModel.cs
@@ -25,12 +25,7 @@ public partial class BarWindowViewModel : ObservableObject
 {
     private static readonly ILogger _log = Log.ForContext("SourceContext", nameof(BarWindowViewModel));
 
-    private const string UnsnapButtonText = "\ue89f";
-    private const string SnapButtonText = "\ue8a0";
-
     private readonly string _errorTitleText = CommonHelper.GetLocalizedString("ToolLaunchErrorTitle");
-    private readonly string _unsnapToolTip = CommonHelper.GetLocalizedString("UnsnapToolTip");
-    private readonly string _snapToolTip = CommonHelper.GetLocalizedString("SnapToolTip");
     private readonly string _expandToolTip = CommonHelper.GetLocalizedString("SwitchToLargeLayoutToolTip");
     private readonly string _collapseToolTip = CommonHelper.GetLocalizedString("SwitchToSmallLayoutToolTip");
 
@@ -38,7 +33,6 @@ public partial class BarWindowViewModel : ObservableObject
     private readonly ExternalToolsHelper _externalToolsHelper;
 
     private readonly ObservableCollection<Button> _externalTools = [];
-    private readonly SnapHelper _snapHelper;
     private readonly DIInsightsService _insightsService;
 
     [ObservableProperty]
@@ -49,15 +43,6 @@ public partial class BarWindowViewModel : ObservableObject
 
     [ObservableProperty]
     private string _systemDiskUsage = string.Empty;
-
-    [ObservableProperty]
-    private bool _isSnappingEnabled = false;
-
-    [ObservableProperty]
-    private string _currentSnapButtonText = SnapButtonText;
-
-    [ObservableProperty]
-    private string _currentSnapToolTip;
 
     [ObservableProperty]
     private string _currentExpandToolTip;
@@ -113,7 +98,6 @@ public partial class BarWindowViewModel : ObservableObject
     {
         _dispatcher = Microsoft.UI.Dispatching.DispatcherQueue.GetForCurrentThread();
 
-        IsSnappingEnabled = TargetAppData.Instance.HWnd != HWND.Null;
         TargetAppData.Instance.PropertyChanged += TargetApp_PropertyChanged;
 
         PerfCounters.Instance.PropertyChanged += PerfCounterHelper_PropertyChanged;
@@ -136,10 +120,7 @@ public partial class BarWindowViewModel : ObservableObject
             ApplicationHwnd = TargetAppData.Instance.HWnd;
         }
 
-        CurrentSnapButtonText = IsSnapped ? UnsnapButtonText : SnapButtonText;
-        CurrentSnapToolTip = IsSnapped ? _unsnapToolTip : _snapToolTip;
         CurrentExpandToolTip = ShowingExpandedContent ? _collapseToolTip : _expandToolTip;
-        _snapHelper = new();
 
         _externalToolsHelper = Application.Current.GetService<ExternalToolsHelper>();
         ((INotifyCollectionChanged)_externalToolsHelper.FilteredExternalTools).CollectionChanged += FilteredExternalTools_CollectionChanged;
@@ -160,91 +141,10 @@ public partial class BarWindowViewModel : ObservableObject
         ExternalToolSeparatorVisibility = _externalToolsHelper.FilteredExternalTools.Count > 0 ? Visibility.Visible : Visibility.Collapsed;
     }
 
-    partial void OnIsSnappedChanged(bool value)
-    {
-        CurrentSnapButtonText = IsSnapped ? UnsnapButtonText : SnapButtonText;
-        CurrentSnapToolTip = IsSnapped ? _unsnapToolTip : _snapToolTip;
-    }
-
-    partial void OnBarOrientationChanged(Orientation value)
-    {
-        if (value == Orientation.Horizontal)
-        {
-            // If we were snapped, unsnap
-            UnsnapBarWindow();
-        }
-        else
-        {
-            // Don't show expanded content in vertical mode
-            ShowingExpandedContent = false;
-        }
-    }
-
-    public void ResetBarWindowOnTop()
-    {
-        // If we're snapped to a target window, and that window loses and then regains focus,
-        // we need to bring our window to the front also, to be in-sync. Otherwise, we can
-        // end up with the target in the foreground, but our window partially obscured.
-        // We set IsAlwaysOnTop to true to get it in foreground and then set to false,
-        // this ensures we don't steal focus from target window and at the same time
-        // bar window is not partially obscured.
-        IsAlwaysOnTop = true;
-        IsAlwaysOnTop = false;
-    }
-
-    [RelayCommand]
-    public void RotateLayout()
-    {
-        if (BarOrientation == Orientation.Horizontal)
-        {
-            BarOrientation = Orientation.Vertical;
-        }
-        else
-        {
-            BarOrientation = Orientation.Horizontal;
-        }
-    }
-
-    public void SnapBarWindow()
-    {
-        // First need to be in a Vertical layout
-        BarOrientation = Orientation.Vertical;
-        _snapHelper.Snap();
-        IsSnapped = true;
-    }
-
-    public void UnsnapBarWindow()
-    {
-        _snapHelper.Unsnap();
-        IsSnapped = false;
-    }
-
-    [RelayCommand]
-    public void ToggleSnap()
-    {
-        if (IsSnapped)
-        {
-            UnsnapBarWindow();
-        }
-        else
-        {
-            SnapBarWindow();
-        }
-    }
-
     [RelayCommand]
     public void ToggleExpandedContentVisibility()
     {
-        if (!ShowingExpandedContent)
-        {
-            // First need to be in a horizontal layout to show expanded content
-            BarOrientation = Orientation.Horizontal;
-            ShowingExpandedContent = true;
-        }
-        else
-        {
-            ShowingExpandedContent = false;
-        }
+        ShowingExpandedContent = !ShowingExpandedContent;
     }
 
     [RelayCommand]
@@ -269,21 +169,7 @@ public partial class BarWindowViewModel : ObservableObject
 
     private void TargetApp_PropertyChanged(object? sender, PropertyChangedEventArgs e)
     {
-        if (e.PropertyName == nameof(TargetAppData.HWnd))
-        {
-            _dispatcher.TryEnqueue(() =>
-            {
-                IsSnappingEnabled = TargetAppData.Instance.HWnd != HWND.Null;
-
-                // If snapped, retarget to the new window
-                if (IsSnapped)
-                {
-                    _snapHelper.Unsnap();
-                    _snapHelper.Snap();
-                }
-            });
-        }
-        else if (e.PropertyName == nameof(TargetAppData.TargetProcess))
+        if (e.PropertyName == nameof(TargetAppData.TargetProcess))
         {
             var process = TargetAppData.Instance.TargetProcess;
